Copy rendered CRD files dir to /api/bases in make manifests
With the apis sub module, consumer operators just have to import
the apis submodule to get access to TransportURL type definitions.
or envtest it is required to install TransportURL CRD to be able to
use it.

Moving the directory config/crd/bases which holds the generated
files in to the apis module that they are available with the module
and link the dir in the original place does not work as kustomize
autodetect fails on symlinked directories [1].

This updates the `generate` Makefile target to copy config/crd/bases
to /apis on successful generation run.

[1] https://github.com/kubernetes-sigs/kustomize/issues/1886