Check if our go.mod files are in sync
Both the main go.mod and the ./apis/go.mod file points to other
operators. But if the version of those operators are not in sync between
the two go.mod files then we can get strange CRD validation errors when
the openstack-operator is deployed. To avoid that a check is added to
the Makefile and to our github workflows