Add renovatebot config to update openstack deps
This adds a workflow to run the renovate github action
periodically (every 30 minutes).

The renovate config file is configured to update
openstack-k8s-operators as soon as an update is detected.
And once a month it will attempt a general dependency
update for the go.mod.

Additionally, a new pin-custom-bundle-dockerfile.sh is
wired into the Makefiles 'bundle' target to pin
the respective openstack bundle images to the
commit tag referenced in go.mod for that operator.

This should help prevent our artifacts from getting
out of sync with the :latest tagged bundles.

NOTE: Renovate supports a postUpgradeTasks but I choose
note to use them as they require "self hosting".