Skip customized operator bundle image
This patch adds a condition to verify if a operator bundle image
needs to be updated or not.
The following scenario is broken in current version of the script,
and is the workflow that will be used on CI:
1. A PR to *-operator is proposed
2. Job runs and build operator, bundle and catalog images.
3. Images are pushed to CI registry.
4. openstack-operator go.mod is updated with new api reference
5. custom-bundle.Dockerfile is modified to point to CI operator bundle
   image
6. Job builds openstack-operator images, and push to CI registry
7. Deploy steps validates openstack-operator image

When generating openstack-operator bundle, it fails to find the bundle
image on user's quay.io account.

The new condition ignores bundle images that were already modified.