[sample]Deploy nova with collapsed cells
In preparation to switch nova to deploy a proper cell setup by default
this patch changes the default OpenStackControlPlane sample in this repo
to use a collapsed cell setup. The collapsed cell setup uses a single
RabbitMQCluster for both the top level nova services and the cell1
services and disables the nova-conductor service in cell0. The cell1
conductor will have access to the API DB and use the common RabbitMQ
so it will act both as a super-conductor and as the cell1 conductor.
This is the same how tripleoo deploys nova by default. However this
means no real cell separation.

To allow NovaConductor to reach Ready in cell0 with the collapsed cell
setup we need the fix https://github.com/openstack-k8s-operators/nova-operator/pull/241
so this patch bump the nova-operator/api dependency version to the
latest one that has that fix.

This patch also adds a new sample that configures OpenStackControlPlane
with two RabbitMQClusters and configure nova that the top level services
uses the common message bus while cell1 use a dedicated message bus to
achieve cell separation.