Don't set status or reconcile disabled services
We have some optional services that don't necessarily be enabled
to deploy a cluster. For example Cinder. Currently, we will try to
update the status even if they aren't enabled.

This ultimately prevents us from reporting a successful deployment. As
such, this patch ensures that the statuses of these objects
are only reported if the service is indeed enabled. For services that
are disabled, we handle them by marking as Ready, with a message to
indicate they have been disabled.