Allows to pass base container as build-arg
At least, using buildah 1.28.0, it seems to be stable enough to pass the
images as proper build-arg, allowing to provide a clean interface to
override the ones we want.

With this patch exposing all of the images as build arguments, we can
even override the golang base container.

Proper usage is, for instance:
```
buildah bud -t my-test:latest \
        --build-arg KEYSTONE_BUNDLE=custom.local/openstack-k8s/keystone-operator-bundle:latest .
```