Use cell separation by default
As the intention is to use separate rabbitmq for each nova cells by
default this patch switches the default OpenStackControlPlane sample to
do so.

Note that this means one extra rabbitmq service is deployed. So it
increases the resource usage of the deployment. I checked with CRC
 $ crc version
 CRC version: 2.13.1+b839ca8
 OpenShift version: 4.12.0
 Podman version: 4.3.1

Deploying the default sample is possible with a CRC VM with 16G RAM. If
you change the rabbitmq resource limits from 2Gi to 1Gi (or even 500Mi)
then I believe it is possible to deploy the default sample with 14G RAM.