Configure to test webhooks locally (#211)
This target will generate self-signed certs and configure the
webhook for it to be tested locally with 'make run-with-webhook'.

You can use $SKIP_CERT to disable cert generation in subsequent
runs and use the exsiting certificates. But it's possible that
the certs have expired/not there and you get an error.