Remove omitempty from <ServiceOperator>.enabled
In the following definition

type <Service> struct {
	// +kubebuilder:validation:Optional
	// +kubebuilder:default=true
	// Enabled - Whether <Service> service should be deployed and managed
	Enabled bool `json:"enabled,omitempty"`

omitempty causes that if the enabled is set to false then it is
serialized to json then the json output will not have an enabled field
as false is the empty golang value for the bool type. Then if that json
is deserialized then the missing enabled field will be defaulted to true
based on the kubebuilder:default. This means the above definition does
not allow setting enabled to false.

Setting enable to false worked before so I think the turning on of the
webhook triggers the extra deserialization / serialization step in the
input processing causing the issue.

This patch removes the omitempty from the enabled fields to prevent the
issue.