Cleanup delete code path
The PR #205 simplified the deletion code path. Now the
OpenStackControlPlane CR deletion is fully automatic as there is no
Finalizer on it. So the reconciler will get an NotFound when called to
reconcile a deleted instance instead of seeing the instance with delete
timestamp set. So that dead code is removed and logging in moved from it
to the NotFound handler so that we still log something when the resource
is deleted.