Split out openstack-storage-bundle
This patch splits out the openstack-storage-bundle. We may take a
similar approach for grouping other operators to save space, but
limit the number of bundle packages we maintain in the catalog.

The csv-merger gets 3 new options:
 --import-env-files Comma separated list of file names to read default operator ENVs from. Used for inter-bundle ENV merging.
 -- export-env-file: Name the external file to write operator ENVs to. Used for inter-bundle ENV merging.
 --alm-examples: a boolean to indicate whether almExamples should be
 processed for merged CSV (false by default we now only keep these in the main openstack-operator/bundle). This save space and we really only need them for top level CRDs. (TODO: are we going to move Dataplane into openstack-operator?, if not may need custom way to handle this)

With the above changes. The build process now works like this:
 -build openstack-operator-storage-bundle (--export-env-file is set
 where we write ENV vars to merge)
 -build openstack-operator-bundle (--import-env-files is used to read in
 the ENV vars from the storage bundle above)
 -a dependency.yaml now installs openstack-storage-operators when
 openstack-operator is installed

If we like this process further updates can be made to split out Rabbit,
Network, and Ironic.

Rename storage-bundle-build-push to dep-bundle-build-push

Add docs on bundle, index build order