Default service specs even if the service is disabled
Even if the service is disabled the service spec needs to be valid
according to then schema so that it can be stored in etcd.

If there is a field like NovaSpec.NovaMetadataTemplate.Enabled that is
defined as *bool but the default value of it is only provided by the
NovaSpec defaulting webhook (as the default value is context dependent)
then if the defaulting is not run on the NovaSpec then the field value
remains null, and that fails the schema validation.

So this patch make the defaulting unconditional for every service spec.