Add new glanceAPI struct
This change is the first of a series where the purpose is to move away
from the harcoded internal/external glanceAPI instance in favor of an
arbitrary number of GlanceAPI instances.
The current change doesn't introduce the ability to deploy >1 GlanceAPI
instance, but introduces the logic that will serve this purpose.
The split behavior is still preserved, but it happens behind the scenes
in Glance.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>