[glance] control route create based on annotation
ATM the glance pkg uses the KeystoneEndpoint parameter to check
if a route for which glance-api should get created:

https://github.com/openstack-k8s-operators/openstack-operator/blob/f36c6c4fe938ffe084e5cd556122f1ba8dfe88ed/pkg/openstack/glance.go#L84

This works, but results when the api to expose and register in keystone
gets changed, the existing route for the previous set glance-api
won't be deleted.

This changes the glance pkg to use the intended annotation on the
glance-api services to create the route for the glance-api services
with "core.openstack.org/ingress_create": "true". The switched api
will have the "core.openstack.org/ingress_create": "false" and if
there is an existing route it gets deleted.

Depends-On: https://github.com/openstack-k8s-operators/glance-operator/pull/401