Prevent endpoint update if service operator changes condition
Previously to apply service overrides to service CRs there
were checks that services have reached a specific condition
state. E.g. for glance that glancev1.GlanceAPIReadyCondition
is true. This conditon resulted in the behavior that when
after a successful deployment glance flipped to
glancev1.GlanceAPIReadyCondition == False e.g. seen when ceph
is not healty and glance api pods manually restarted, that the
service overrides got removed (with route hostname information)
and the keystoneendpoint updated to be the k8s service instead
of the route.

With this change there is no wait for a condition, instead
checking the endpoint k8s services matching the label selector
applied by the openstack-operator. If no services are found, the user
provided overrides + the label selector get set. If the expected
number of k8s services are identified, endpointdetails get
evaluated and e.g. the routes hostname applied.

Initially the conditions were used to prevent rerunning bootstrap
jobs while there is still one in progress. With the changes
to the DoJob [1] this should no longer be required since the job
handling will wait for a running job to complete before rerunning
it with a new config.

[1] https://github.com/openstack-k8s-operators/lib-common/blob/main/modules/common/job/job.go#L115