[kuttl] change cellDatabaseUser to in the collapsed cell test
The cellDatabaseUser field is replaced with the cellDatabaseAccount.
Most of the sample and kuttl test assert files does not define the
cellTemplates and therefore they are defaulted by nova-operator. However
in the collapsed cell sample we need to finetune the cell template
fields and therefore some of the cellTemplate fields are explicitly
defined. However if at least one of the field is defined then the struct
level defaulting does not triggered, so the rest of the fields needs to
be define too, otherwise we get the golang default values.

This patch makes sure that cellDatabaseAccount field is defined in the
samples where we explicitly define cellTemplates.