[tls] fix status.tls.caList[].expires
currently always the first element of the bundle list gets used
to get the expire date of the cacert to be added to the caList in
the status. This is always the same cert and the expire dates will
show the same for all ca certs:

  tls:
    caBundleSecretName: combined-ca-bundle
    caList:
    - expires: "2029-03-14T14:26:31Z"
      name: rootca-public
    - expires: "2029-03-14T14:26:31Z"
      name: rootca-internal
    - expires: "2029-03-14T14:26:31Z"
      name: rootca-ovn

With the fix a temporary bundle is used to get the details for the
current cacert. Now the expires time stamps are as expected:

  tls:
    caBundleSecretName: combined-ca-bundle
    caList:
    - expires: "2029-03-14T15:19:05Z"
      name: rootca-public-custom
    - expires: "2024-07-01T14:33:01Z"
      name: rootca-internal
    - expires: "2024-07-01T14:22:37Z"
      name: rootca-ovn