Version: Create helper before manipulating instance
The helper.NewHelper creates snapshot of the instance being reconciled
and PatchInstance at the end of the reconcile loop uses this to generate
the Patch to persist the changes done by the reconciler. Therefore the
helper needs to be created before any changes on the instance. This is
fixed now to avoid loosing in memory changes.