Enable TLS for glanceAPI Edge instances
When a glanceAPI Edge instance is deployed, the glance-operator creates
an internal SVC, but the public one is skipped (because these instances
are not exposed outside).
However, we're currently not able to enable TLS because the EnsureEndpoint
function is skipped due to the different size of the generated Overrides
that is compared with the service list (svcs.Items).
This patch fixes the current behavior and generates an internal override
only for an Edge API instance type: by doing this we're able to properly
execute the EnsureEndpoint function and generate both key and cert,
referenced by the Edge API.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>