Additional SubjectName in the CertificateRequest
This patch adds a way to check the appropriate annotation within a SVC
and look for additional SNs that should be added to the
CertificateRequest. Glance needs this mechanism because for each SVC
(public, internal), we have an associated headless service, and it is
used to resolve each replica Pod worker_self_reference_url.
This allows to proxy Pod2Pod requests via HTTPS.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>