Move service templates to be ptr to struct
Moves template parameter in the service sections to be a ptr to
be them omitted in case its not provided.
Also only triggers the defaulting for the service in the
webhook if the service is enabled or a template for the service
was provided.

If the service was disabled and gets enabled, at least an empty
`template: {}` needs to be passed for correct defaulting to
happen.

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>