update config samples to include default cell template
This change updates the config samples to inculde the cells templates
so that the default behavior fo the cell 1 creation can be modifed in
the future without impacting docs or ci jobs.

currently cell 0 and 1 are created by nova by default
in a feture operator version we plannign to remove the automatic creation
of cell 1 so that the procedure ot add the first and subsequent comptue
cells are identical.