[envtest]Stabilize version controller tests
There was a race condition between the version controller applying the
finalizer to the version instance and the test case cleanup removing
such finalizer. By waiting for the controller to run at least once
before the test finishes is enough to avoid the race as the controller
only adds the finalizer during the first reconcile run that initialize
the conditions.

Also cleaned up the test case itself a bit and added more precise
assert.

Co-authored-by: Dan Prince <dprince@redhat.com>