Bump Glance with the new storage interface
We currently provide both StorageClass and StorageRequest that can be
propagated to the Glance underlying instances.
However, it might be possible that operators use extraMounts to plug a
NFS share mapped to /var/lib/glance/os_staging_store.
In this case, the API should be updated to not provision any PVC and
let the human operator to take care about it using the extraMounts
interface.
This patch updates the Glance bits to reflect the new behavior
introduced in the glance-operator.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>