Use non static name for the Cinder CR
When the openstack operator creates the cinder CR it always uses the
"cinder" name and there is no way to change it.

This means that the names of the pods, and therefore the reported
hostname inside them, will always be the same for the backup service
(cinder-backup-0). The same will happen for the cinder volume in 2
different deployment if they have the same key in the volumes map.

This may be fine for some deployments, but it creates problems in
others, because many cinder drivers expect the hostnames of the hosts
attaching volumes to be unique.

Because they expect them to be unique they use the hostname as a key to
store information on the storage array such as the iSCSI initiator name
or the HBAs addresses.

If they are not unique there unexpected behavior can happen, like hosts
being unable to attach.

This is very common in CI systems, but is not limited to them as it can
also happen in production deployments that are using the same storage
arrays as their pre-production counterparts (even if it's different
pools).

In this patch the human operator has now a way to ask for unique pod
names in cinder services using the `uniquePodNames` boolean field in
the `cinder` section.

When `uniquePodNames` is set to `true` the openstack-operator will name
the `Cinder` CR using the first 6 characters of the
`OpenStackControlPlane` UID, that way the pods for the cinder services
will have a unique name.

Pod name collision is still possible, but unlikely.

Jira: OSPRH-7396