Use non static name for the Glance CR
When glance uses cinder as a backend it is also affected by the
restrictions of needing unique pod names to have unique hostnames or
risk the possibility of attachments not working as desired.

The solution applied by this patch is the same as applied to cinder. Use
the `uniquePodNames` field to generate a unique `Glance` name.

This allows human operators ensure that the pod names in their
deployment are unique.

Jira: OSPRH-7396