Drop csv-merger
This commit drops the csv-merger as it was no longer in
use for merging CSV's and only used to merge in ENV
variables from the service operators.

Maintain those ENVs in the openstack-operator will mean that each time
an openstack-operator bundle is generated the associated
SHA's for service containers will be pinned to a new version
and thus decoupled from the service operators themselves.