Don't mount on /runner/env/ssh_key
Though this is the default from which ansible-runner uses the
ssh_key, it does not match with `ansible_ssh_private_key_file`
var in the inventrory and it creates issues.

Let's not mount on the default runner location and instead in
location that matches with `ansible_ssh_private_key_file`.

jira: https://issues.redhat.com/browse/OSPRH-8085
Signed-off-by: rabi <ramishra@redhat.com>