Create cloudrc file in `openstackclient` pod
We are currently only creating the `clouds.yaml` and `secure.yaml` in
the `openstackclient` pod, and that is enough for the openstack client,
but unfortunately this is not enough since there is no parity between
that client and the individual clients.

For example it is missing (though it has upstream patches in progress):

- The `openstack volume service list` cannot show the cluster.
- The `openstack volume service list` cannot show the backend state.
- There is no support for any of the default volume type commands: set,
  show, list, unset.

These are some of the ones we know about, but there could still be some
other missing.

Users will need to use the old `cinder` client, so we should make it
convenient for them by having the `cloudrc` available in the
`openstackclient` pod.

Depends-On: https://github.com/openstack-k8s-operators/keystone-operator/pull/432