Clean up apiOverrides when a GlanceAPI is removed from ctlplane
Currently, if a GlanceAPI is added to the ctlplane and then removed, the
human operator needs to manually clean up glance section in the ctlplane
CR and remove the generated APIOverrides.
In addition to that, overrides are generated regardless of the nature of
the glanceAPI (e.g. Edge).
This patch solves this problem by removing the overrides for the API
that have been removed and are no longer present in the ctlplane
definition.

Jira: OSPRH-8072

Signed-off-by: Francesco Pantano <fpantano@redhat.com>