Consolidate Ansibleee controller logic
This change consolidates the openstack-ansibleee-operator logic into the
openstack-operator.
The openstack-ansibleee-operator currently is a full operator just to
create Kubernetes jobs. It would be better if this functionality was
just included with the rest of the dataplane code. This change keeps the
controller, but runs it under the openstack-operator instead of
standalone.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>
Jira: https://issues.redhat.com/browse/OSPRH-8718