Add edpm_service_types
The list of service types is added as an ansible variable to the
inventory as edpm_service_types. This will be used instead of
edpm_services since service names could be any custom values.

edpm-ansible roles such as edpm_update, need a list of consistent
service identifiers (types instead of names), so it knows which other
roles to include for update tasks.

Jira: https://issues.redhat.com/browse/OSPRH-8107
Signed-off-by: James Slagle <jslagle@redhat.com>