Revert "Remove the deployed attribute from deployment"
This reverts commit 5a55c47a24fca6ac76ca08c72f9a0c69913379d9.

Without the status.Deployed boolean flag on OpenStackDataPlaneDeployment
there is no mechanism in place to prevert the full reconcile of a
Deployment. While the immutable annotation prevents user edits to the
spec, other changes could still trigger Deployment reconciles.

In the case of testing minor updates, it was discovered that when the
ansible runner image as specified by the
RELATED_IMAGE_ANSIBLEEE_IMAGE_URL_DEFAULT env var in the
openstack-operator csv is changed, it will re-execute the ansible for
all existing Deployments. This breaks the minor update workflow as
things are done out of order, and breaks the expectation that it is the
user who starts ansible executions as opposed to automatically.

Jira: OSPRH-8923
Signed-off-by: James Slagle <jslagle@redhat.com>