Log when *ImageMatch functions return false
This adds Log.Info statements when the images
do not match up which can help pin-point debugging.

If the service contains a single image both images
are logged if there is a mismatch. To keep code
more readable for multi-container services only
a simple log message is added in the case of a
mismatch (still useful enough to pin-point
an issue).