Default to global spec.secret like other service do
Most of the service default to spec.secret if the service.spec.secret
is not provided. This adds it to those services which not yet did that.

This should probably be moved to be done in the webhook, but for now
this gets it in sync with the current implementation of the other
services

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>