Use empty Alm examples
Previously with csv-merger we removed the Alm examples
in the CSV. This is to save space as our bundles are quite large.

Now that we are using the raw data generated by operator-sdk's
'generate bundle' command the config/sample data is getting merged into the
generated CSV for openstack-operator. As we had multiple controlplanes
in the kustomization.yaml file this meant lots of extra YAML examples
which aren't needed.

In order to maintain the previous behavior of effectively
removing ALM examples, but still passing bundle generate
validations this commit creates '_empty.yaml' versions of
each CR and then uses those in the kustomization.yaml.