Unmarshal values from AnsibleVarsFrom as json
Instead of assuming that the ansible vars in a configmap or secret used
with AnsibleVarsFrom are just string key/value pairs, support json
values.  This is needed to support variables such as
edpm_container_registry_logins with AnsibleVarsFrom.

To maintain backwards compatiblity with any already created ConfigMaps
or Secrets, if the unmarshalling to JSON fails, a plain string value is
assumed.

Jira: https://issues.redhat.com/browse/OSPRH-9500
Signed-off-by: James Slagle <jslagle@redhat.com>