OpenStackVersion: Add releaseVersionScheme and csvVersionAppend support
The releaseVersionScheme can be optionally used to enable 'csvEpochAppend' behavior
This is used by some downstream builds to append the version to the CSV version suffix (epoch) which
is obtained from the OPERATOR_CONDITION_NAME environment variable to the openstack release version.
In some downstream build systems CSV version bumps can be easily automated where as the
OPENSTACK_RELEASE_VERSION is more of a static constant.

Co-authored-by: Miguel Garcia(mgarciac@redhat.com)

Jira: OSPRH-9568