Use OCP registries.conf for disconnected deployments
This change adds functionality to check for ImageContentSourcePolicy in the OCP cluster.
If we find it, we infer from it's existance that we are running in a disconnected OCP
deployment. If this is indeed the case, we will pull the registries.conf file
from the MachineConfig and use this for our dataplane nodes configured via
edpm_ansible.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>

Update RBAC for ICSP and MachineConfig

This change allows the OpenStackDataPlaneNodeSet controller to get and list
ImageContentSourcePolicy and MachineConfig CR's from the cluster. This facilitates
integration of the dataplane with the OCP registries.conf configuration to ensure
consistency.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>

Return err if no prefix

Signed-off-by: Brendan Shephard <bshephar@redhat.com>