Revert "[ovn-controller] Don't create ovn-controller if nicMappings empty"
This reverts commit a02a1596b972a599d77158d9dc746849f8075b23.

This change created a bug on the adoption where all the gates reaches TIME_OUT,
reverting this patch to unblock prod chain while fixing it.

The problem that this patch was introducing was that edpm deployment expects
a config map created by ovn-operator (ovncontroller) called ovncontroller-config.
With this patch the ovncontroller CR was not created and the ovn-operator was
not reconciling this object, which was needed to create the said config map.

Related: OSPRH-7463
Related: OSPCIX-504