Add openstack must-gather to required images
In disconnected environments, we should have a reference to the
openstack-must-gather image. This allows for users to clone in the required image to their
disconnected registries and enable them to gather logs for support cases and troubleshooting.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>
(cherry picked from commit 6cac6382dac5a69bf3a93a4fa88024bc23823b30)