Update Samples and configure Swift backend for Glance
The unsupported File backend is currently used in most of the samples.
They are not consistent even with the documentation, where we suggest to
add replicas: 0 to the glance service until a backend is chosen.
Considering we have Swift deployed by default, this patch fixes the
Glance samples and add Swift as a backend. This change will be
reflected in the openstack-operator guide as well.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>