Switch EDPMServiceType to Enum
This change switches the EDPMServiceType parameter to be an Enum
that accepts a fixed list of service names. This is required, since allowing
arbitrary names can interfere with the TLS mounts desribed in the referenced
Jira.

Instead of defaulting to the name of the service, we will instead
default to generic and attempt to mount in a generic CA cert to ensure
functionality of the dataplane.

Jira: https://issues.redhat.com/browse/OSPRH-10804
Signed-off-by: Brendan Shephard <bshephar@redhat.com>